github_stats/requirements.txt
boto3==1.17.0
mock==4.0.3
moto==1.3.16
pytest==6.2.2
PyGitHub==1.54.1
requests==2.24.0
